agent-runtime/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
redis==5.0.1
qdrant-client==1.7.0
langchain==0.1.0
langchain-openai==0.0.2
PyGithub==2.1.1
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
